requirements.txt
aiocron~=1.8
asgiref~=3.6.0
Django~=4.1.5
folium~=0.14.0
freezegun~=1.2.2
openai~=0.27.2
openpyxl~=3.0.10
Pillow~=9.4.0
pytz~=2022.7.1
requests~=2.28.2
shapely~=2.0.0
XlsxWriter~=3.0.7
pydub~=0.25.1


asynctest~=0.13.0